+++ src/app/PubSubVitessceGrid.js
@@ -2,6 +2,7 @@ import React from 'react';
 import VitessceGrid from 'vitessce-grid';
 
 import { LayerPublisher } from '../components/layerpublisher';
+import { getPreferredTheme } from '../components/utils';
 
 export default class PubSubVitessceGrid extends React.Component {
   constructor(props) {
@@ -13,8 +14,9 @@ export default class PubSubVitessceGrid extends React.Component {
   render() {
     const { config, getComponent } = this.props;
     const { allReady } = this.state;
+    const theme = config.theme || getPreferredTheme();
     return (
-      <div className="vitessce-container">
+      <div className={`vitessce-container vitessce-theme-${theme}`}>
         { allReady && <LayerPublisher layers={config.layers} /> }
         <VitessceGrid
           layout={config.staticLayout}